Add cast to velocity setting to avoid warning.

+++ Star.cpp
@@ -15,7 +15,7 @@
 Star::Star() {
   setType("Star");
   setSolidness(df::SPECTRAL);
-  setXVelocity(-1.0 / (rand()%10 + 1));
+  setXVelocity((float) (-1.0 / (rand()%10 + 1)));
   setAltitude(0);	// Make them in the background.
   df::GraphicsManager &graphics_manager = df::GraphicsManager::getInstance();
   df::Position pos(rand()%graphics_manager.getHorizontal(),
@@ -47,5 +47,5 @@ void Star::out() {
   df::Position pos(graphics_manager.getHorizontal() + rand()%20,
 	       rand() % graphics_manager.getVertical());
   setPosition(pos);
-  setXVelocity(-1.0 / (rand()%10 + 1));
+  setXVelocity((float)(-1.0 / (rand() % 10 + 1)));
 }
